+++ hypex/extensions/scipy_linalg.py
